making suggested changes
